chore: Update npm dependencies and scripts
This commit updates the `nodemon` dependency to the latest stable version (`3.1.3`) in the `devDependencies` section of the `package.json` file. It also updates the `start` script in the `scripts` section to use the updated `nodemon` version.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a brief description of the changes.